chore: saving work before qa phase
Split cli.test.ts into 7 domain-specific test files with shared helper.
All 175 cli tests pass, full monorepo green.
